+++ src/actions/create_issue.py
@@ -397,13 +397,8 @@ class UserOutput(ActionOutput):
 
 
 class CreateIssueOutput(PermissiveActionOutput):
-    # PermissiveActionOutput so that every field GitHub returns for an issue is
-    # passed through to the client (playbooks may key off fields we don't model,
-    # e.g. reactions, pull_request, state_reason). Declared fields drive the
-    # widget columns and CEF pivots; unknown fields flow through instead of
-    # being dropped.
     # Column fields in widget display order
-    number: float = OutputField(
+    number: int = OutputField(
         cef_types=["github issue id"], example_values=[2], column_name="Issue Number"
     )
     title: str = OutputField(
@@ -464,7 +459,7 @@ class CreateIssueOutput(PermissiveActionOutput):
 
 
 class CreateIssueSummary(ActionOutput):
-    issue_number: float | None = OutputField(example_values=[1])
+    issue_number: int | None = OutputField(example_values=[1])
     issue_url: str | None = OutputField(
         cef_types=["url"], example_values=["https://github.com/test/test-repo/issues/1"]
     )
