+++ src/actions/get_issue.py
@@ -122,7 +122,7 @@ class GetIssueParams(Params):
     repo_name: str = Param(
         description="Name of the repository", primary=True, cef_types=["github repo"]
     )
-    issue_number: float = Param(
+    issue_number: int = Param(
         description="Issue ID", primary=True, cef_types=["github issue id"]
     )
 
@@ -373,11 +373,6 @@ class ClosedByOutput(ActionOutput):
 
 
 class GetIssueOutput(PermissiveActionOutput):
-    # PermissiveActionOutput so that every field GitHub returns for an issue is
-    # passed through to the client (playbooks may key off fields we don't model,
-    # e.g. labels, reactions, pull_request, state_reason). Declared fields drive
-    # the datapaths/CEF pivots; unknown fields flow through instead of being
-    # dropped.
     assignee: AssigneeOutput | None
     assignees: list[AssigneesOutput]
     author_association: str = OutputField(example_values=["OWNER"])
@@ -416,7 +411,7 @@ class GetIssueOutput(PermissiveActionOutput):
     node_id: str = OutputField(
         example_values=["MDU6SXNzdWU0Njg4MzQwOTA="]  # pragma: allowlist secret
     )
-    number: float = OutputField(cef_types=["github issue id"], example_values=[1])
+    number: int = OutputField(cef_types=["github issue id"], example_values=[1])
     repository_url: str = OutputField(
         cef_types=["url"],
         example_values=["https://api.github.com/repos/repoowner/TestingAPI"],
@@ -432,7 +427,7 @@ class GetIssueOutput(PermissiveActionOutput):
 
 
 class GetIssueSummary(ActionOutput):
-    issue_number: float | None = OutputField(example_values=[1])
+    issue_number: int | None = OutputField(example_values=[1])
     issue_url: str | None = OutputField(
         cef_types=["url"], example_values=["https://github.com/test/test-repo/issues/1"]
     )
@@ -442,7 +437,7 @@ def get_issue(params: GetIssueParams, soar: SOARClient, asset: Asset) -> GetIssu
     endpoint = GITHUB_ENDPOINT_GET_ISSUE.format(
         repo_owner=params.repo_owner,
         repo_name=params.repo_name,
-        issue_number=int(params.issue_number),
+        issue_number=params.issue_number,
     )
     response = call_github("GET", endpoint, asset)
     _check_response(response)
